backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
pyodbc==5.2.0
pydantic-settings==2.10.1
pandas==2.3.2
arkofdata-common @ git+https://github.com/wayneweicheng/utilities_arkofdata-common.git@v0.1.0
psutil==6.0.0
pywinauto==0.6.8
ib_insync==0.9.86
langchain-core>=0.3.37
langchain-openai>=0.2.16
email-validator>=2.0.0
